+++ src/app/Spotify/services/spotify-player.service.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-import { HttpService } from '../http.service';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
